requirements.txt
openai
python-dotenv
matplotlib
python-docx
Pillow
